[REF] odoo: remove test code from registry and sql_db
There is no need to have `TestCursor` or test related methods in non
test modules as such `TestCursor` has been moved to its own file and
test related methods in registry have been removed in favor of patching
during tests.

`odoo.sql_db.TestCursor` -> `odoo.tests.test_cursor.TestCursor`
`odoo.orm.registry.enter_test_mode` ->
`odoo.tests.TransactionCase.registry_enter_test_mode` (same for leave)

This will allow further control on what we can and can not do in the
`odoo.tests` module.
